fix(lint): resolve consistent-return, no-unstable-nested-components, and import cycle in VisualizationBlock
